requirements.txt
watchdog==2.1.6
mysql-connector-python==8.0.31
requests==2.31.0
importlib-metadata==6.0.0
